Move contexts into own files
