+++ sandbox/grommet-app/src/contexts/BackgroundContext.jsx
@@ -1,0 +1,3 @@
+import { createContext } from 'react';
+
+export const BackgroundContext = createContext({});
