+++ sandbox/grommet-app/src/contexts/WorkspaceContext.jsx
@@ -1,0 +1,3 @@
+import { createContext } from 'react';
+
+export const WorkspaceContext = createContext({});
