+++ sandbox/grommet-app/src/contexts/index.js
@@ -1,0 +1,2 @@
+export { BackgroundContext } from './BackgroundContext';
+export { WorkspaceContext } from './WorkspaceContext';
